+++ keyring/src/service.rs
@@ -1,26 +1,42 @@
 //! # Keyring service name
 //!
-//! Module dedicated to global service name management. Every consumer
-//! of this crate should define his own service name.
+//! Module dedicated to global keyring service name management. Every
+//! consumer of this crate should define his own service name at the
+//! beginning of their program.
 
 use log::debug;
 use once_cell::sync::OnceCell;
 
 /// The global service name, wrapped in a once cell.
-static SERVICE_NAME: OnceCell<&str> = OnceCell::with_value("keyring-lib");
+static SERVICE_NAME: OnceCell<&str> = OnceCell::new();
+
+/// The default global service name.
+static DEFAULT_SERVICE_NAME: &str = "keyring-lib";
 
 /// Get the global keyring service name.
 ///
-/// The service name is used every time a new native entry is created.
+/// If the service name is not defined, returns the default global
+/// service name `keyring-lib`.
 pub fn get_global_service_name() -> &'static str {
-    SERVICE_NAME.get().unwrap()
+    match SERVICE_NAME.get() {
+        Some(name) => name,
+        None => {
+            let err = format!("service name not defined, defaults to `{DEFAULT_SERVICE_NAME}`");
+            debug!("cannot get global keyring service name: {err}");
+            DEFAULT_SERVICE_NAME
+        }
+    }
 }
 
 /// Replace the global keyring service name.
-pub fn set_global_service_name(next: &'static str) {
-    if let Err(prev) = SERVICE_NAME.set(next) {
-        debug!("global keyring service name {prev} replaced by {next}");
-    } else {
-        debug!("global keyring service name set to {next}");
+///
+/// This action as no effect if a global service name has already been
+/// defined.
+pub fn set_global_service_name(name: &'static str) {
+    debug!("setting global keyring service name `{name}`");
+
+    if let Err((prev, _)) = SERVICE_NAME.try_insert(name) {
+        let err = format!("service already named `{prev}`");
+        debug!("cannot set `{name}` as global keyring service name: {err}");
     }
 }